requirements.txt
Django==2.0.6
django-mathfilters==0.4.0
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
requests==2.18.4
celery==4.2.0
django-encrypted-model-fields==0.5.5
django-heroku==0.3.1
django-grappelli==2.11.1
gunicorn==19.9.0
attrs==18.1.0
